# api/web_page/openapi.yaml
swagger: "2.0"
info:
    title: Mai.Today Server
    description: Mai.Today Server
    version: 1.0.0
host: dev.api.mai.today
consumes:
    - application/json
    - application/xml
    - application/gob
produces:
    - application/json
    - application/xml
    - application/gob
paths:
...
    /{channel}/sync:
        get:
            tags:
                - Sync
            summary: '[WebSocket] 請求同步'
            operationId: Sync#receiveSync
            parameters:
                - name: channel
                  in: path
                  description: 使用者識別碼
                  required: true
                  type: string
                  maxLength: 24
                  minLength: 24
                - name: Authorization
                  in: header
                  required: false
                  type: string
            responses:
                "101":
                    description: 此 API 僅用於 WebSocket 協議
                "200":
                    description: 操作成功
                    schema:
                        $ref: '#/definitions/SyncReceiveSyncOKResponseBody'
                        required:
                            - command
                            - timestamp
                            - data
            schemes:
                - https
            security:
                - JWT_header_Authorization: []
...
    /sync:
        post:
            tags:
                - Sync
            summary: 請求同步
            operationId: Sync#sync
            parameters:
                - name: Authorization
                  in: header
                  required: false
                  type: string
                - name: SyncRequestBody
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/SyncSyncRequestBody'
            responses:
                "200":
                    description: 操作成功
                    schema:
                        $ref: '#/definitions/SyncSyncOKResponseBody'
                        required:
                            - command
                            - timestamp
                            - data
                "401":
                    description: Unauthorized response.
                    schema:
                        $ref: '#/definitions/SyncSyncInvalidTokenResponseBody'
                "500":
                    description: 伺服器發生錯誤
            schemes:
                - https
            security:
                - JWT_header_Authorization: []
definitions:
...
    BaseInfoResponseBody:
        title: BaseInfoResponseBody
        type: object
        properties:
            baseId:
                type: string
                description: 基地識別碼
                example: 669071b8a650a662b82285ca
                minLength: 24
                maxLength: 24
            color:
                type: string
                description: 主題颜色
                example: '#FF5733'
                minLength: 1
                maxLength: 16
            deleted_at:
                type: string
                description: 刪除時間
                example: "2024-01-01T11:53:54Z"
                format: date-time
            id:
                type: string
                description: 識別碼
                example: 669071b8a650a662b82285ca
                minLength: 24
                maxLength: 24
            logo:
                type: string
                description: 標識圖片的 URL
                example: http://example.com/logo.png
                minLength: 1
                maxLength: 1024
            name:
                type: string
                description: 名稱
                example: 基地 A
                minLength: 1
                maxLength: 128
        description: 基地資訊
        example:
            baseId: EbZvovBKbFogJeRem5oj2S
            color: '#FF5733'
            deleted_at: null
            id: EbZvovBKbFogJeRem5oj2S
            logo: http://example.com/logo.png
            name: 基地 A
        required:
            - id
            - baseId
            - color
            - logo
            - name
    BaseNavStateResponseBody:
        title: BaseNavStateResponseBody
        type: object
        properties:
            baseId:
                type: string
                description: 基地識別碼
                example: 669071b8a650a662b82285ca
                minLength: 24
                maxLength: 24
            deleted_at:
                type: string
                description: 刪除時間
                example: "2024-01-01T11:53:54Z"
                format: date-time
            id:
                type: string
                description: 識別碼
                example: 669071b8a650a662b82285ca
                minLength: 24
                maxLength: 24
            index:
                type: integer
                description: 順序
                default: 1
                example: 0
                format: int64
                minimum: 1
        description: 基地導覽列狀態
        example:
            baseId: 669071b8a650a662b82285ca
            deleted_at: "2024-01-01T11:53:54Z"
            id: 669071b8a650a662b82285ca
            index: 0
        required:
            - id
            - baseId
            - index
...
    CommandResponseBody:
        title: CommandResponseBody
        type: object
        properties:
            type:
                type: string
                description: 類型
                example: updateBaseInfo
                enum:
                    - sync
                    - createBase
                    - deleteBase
                    - reorderBaseNavState
                    - updateBaseInfo
        description: 指令
        example:
            type: reorderBaseNavState
        required:
            - type
...
    SyncPayloadDataResponseBody:
        title: SyncPayloadDataResponseBody
        type: object
        properties:
            baseInfos:
                type: array
                items:
                    $ref: '#/definitions/BaseInfoResponseBody'
                example:
                    - baseId: EbZvovBKbFogJeRem5oj2S
                      color: '#FF5733'
                      deleted_at: null
                      id: EbZvovBKbFogJeRem5oj2S
                      logo: http://example.com/logo.png
                      name: 基地 A
                    - baseId: EbZvovBKbFogJeRem5oj3S
                      color: '#FF5733'
                      deleted_at: null
                      id: EbZvovBKbFogJeRem5oj3S
                      logo: http://example.com/logo.png
                      name: 基地 B
                    - baseId: EbZvovBKbFogJeRem5oj4S
                      color: '#FF5733'
                      deleted_at: null
                      id: EbZvovBKbFogJeRem5oj4S
                      logo: http://example.com/logo.png
                      name: 基地 C
            baseNavStates:
                type: array
                items:
                    $ref: '#/definitions/BaseNavStateResponseBody'
                example:
                    - baseId: EbZvovBKbFogJeRem5oj2S
                      deleted_at: null
                      id: EbZvovBKbFogJeRem5oj7a
                      index: 0
                    - baseId: EbZvovBKbFogJeRem5oj3S
                      deleted_at: null
                      id: EbZvovBKbFogJeRem5oj7b
                      index: 1
                    - baseId: EbZvovBKbFogJeRem5oj4S
                      deleted_at: null
                      id: EbZvovBKbFogJeRem5oj7c
                      index: 2
        example:
            baseInfos:
                - baseId: EbZvovBKbFogJeRem5oj2S
                  color: '#FF5733'
                  deleted_at: null
                  id: EbZvovBKbFogJeRem5oj2S
                  logo: http://example.com/logo.png
                  name: 基地 A
                - baseId: EbZvovBKbFogJeRem5oj3S
                  color: '#FF5733'
                  deleted_at: null
                  id: EbZvovBKbFogJeRem5oj3S
                  logo: http://example.com/logo.png
                  name: 基地 B
                - baseId: EbZvovBKbFogJeRem5oj4S
                  color: '#FF5733'
                  deleted_at: null
                  id: EbZvovBKbFogJeRem5oj4S
                  logo: http://example.com/logo.png
                  name: 基地 C
            baseNavStates:
                - baseId: EbZvovBKbFogJeRem5oj2S
                  deleted_at: null
                  id: EbZvovBKbFogJeRem5oj7a
                  index: 0
                - baseId: EbZvovBKbFogJeRem5oj3S
                  deleted_at: null
                  id: EbZvovBKbFogJeRem5oj7b
                  index: 1
                - baseId: EbZvovBKbFogJeRem5oj4S
                  deleted_at: null
                  id: EbZvovBKbFogJeRem5oj7c
                  index: 2
    SyncReceiveSyncOKResponseBody:
        title: SyncReceiveSyncOKResponseBody
        type: object
        properties:
            command:
                $ref: '#/definitions/CommandResponseBody'
            data:
                $ref: '#/definitions/SyncPayloadDataResponseBody'
            timestamp:
                type: integer
                description: 時間戳記
                example: 1719188007000000000
                format: int64
                minimum: 1.719188007e+18
        example:
            command:
                type: sync
            data:
                baseInfos:
                    - baseId: EbZvovBKbFogJeRem5oj2S
                      color: '#FF5733'
                      deleted_at: null
                      id: EbZvovBKbFogJeRem5oj2S
                      logo: http://example.com/logo.png
                      name: 基地 A
                    - baseId: EbZvovBKbFogJeRem5oj3S
                      color: '#FF5733'
                      deleted_at: null
                      id: EbZvovBKbFogJeRem5oj3S
                      logo: http://example.com/logo.png
                      name: 基地 B
                    - baseId: EbZvovBKbFogJeRem5oj4S
                      color: '#FF5733'
                      deleted_at: null
                      id: EbZvovBKbFogJeRem5oj4S
                      logo: http://example.com/logo.png
                      name: 基地 C
                baseNavStates:
                    - baseId: EbZvovBKbFogJeRem5oj2S
                      deleted_at: null
                      id: EbZvovBKbFogJeRem5oj7a
                      index: 0
                    - baseId: EbZvovBKbFogJeRem5oj3S
                      deleted_at: null
                      id: EbZvovBKbFogJeRem5oj7b
                      index: 1
                    - baseId: EbZvovBKbFogJeRem5oj4S
                      deleted_at: null
                      id: EbZvovBKbFogJeRem5oj7c
                      index: 2
            timestamp: 1719188007000000000
        required:
            - command
            - timestamp
            - data
    SyncSyncInvalidTokenResponseBody:
        title: 'Mediatype identifier: application/vnd.goa.error; view=default'
        type: object
        properties:
            fault:
                type: boolean
                description: Is the error a server-side fault?
                example: false
            id:
                type: string
                description: ID is a unique identifier for this particular occurrence of the problem.
                example: 123abc
            message:
                type: string
                description: Message is a human-readable explanation specific to this occurrence of the problem.
                example: parameter 'p' must be an integer
            name:
                type: string
                description: Name is the name of this class of errors.
                example: bad_request
            temporary:
                type: boolean
                description: Is the error temporary?
                example: false
            timeout:
                type: boolean
                description: Is the error a timeout?
                example: false
        description: token 無效 (default view)
        example:
            fault: false
            id: 123abc
            message: parameter 'p' must be an integer
            name: bad_request
            temporary: true
            timeout: true
        required:
            - name
            - id
            - message
            - temporary
            - timeout
            - fault
    SyncSyncOKResponseBody:
        title: SyncSyncOKResponseBody
        type: object
        properties:
            command:
                $ref: '#/definitions/CommandResponseBody'
            data:
                $ref: '#/definitions/SyncPayloadDataResponseBody'
            timestamp:
                type: integer
                description: 時間戳記
                example: 1719188007000000000
                format: int64
                minimum: 1.719188007e+18
        example:
            command:
                type: sync
            data:
                baseInfos:
                    - baseId: EbZvovBKbFogJeRem5oj2S
                      color: '#FF5733'
                      deleted_at: null
                      id: EbZvovBKbFogJeRem5oj2S
                      logo: http://example.com/logo.png
                      name: 基地 A
                    - baseId: EbZvovBKbFogJeRem5oj3S
                      color: '#FF5733'
                      deleted_at: null
                      id: EbZvovBKbFogJeRem5oj3S
                      logo: http://example.com/logo.png
                      name: 基地 B
                    - baseId: EbZvovBKbFogJeRem5oj4S
                      color: '#FF5733'
                      deleted_at: null
                      id: EbZvovBKbFogJeRem5oj4S
                      logo: http://example.com/logo.png
                      name: 基地 C
                baseNavStates:
                    - baseId: EbZvovBKbFogJeRem5oj2S
                      deleted_at: null
                      id: EbZvovBKbFogJeRem5oj7a
                      index: 0
                    - baseId: EbZvovBKbFogJeRem5oj3S
                      deleted_at: null
                      id: EbZvovBKbFogJeRem5oj7b
                      index: 1
                    - baseId: EbZvovBKbFogJeRem5oj4S
                      deleted_at: null
                      id: EbZvovBKbFogJeRem5oj7c
                      index: 2
            timestamp: 1719188007000000000
        required:
            - command
            - timestamp
            - data
    SyncSyncRequestBody:
        title: SyncSyncRequestBody
        type: object
        properties:
            lastTimestamp:
                type: integer
                description: 上次同步的時間戳記
                example: 1719188007000000000
                format: int64
                minimum: 1.719188007e+18
        example:
            lastTimestamp: 1719188007000000000
securityDefinitions:
    JWT_header_Authorization:
        type: apiKey
        name: Authorization
        in: header